readMe.md
We're going to create a function that takes the string we want to find the longest word of as an argument. Then we are going to turn that string into an array by using split() to turn each individual word into its own array index by splitting them by whitespace.

Next we're going to run a for loop to cycle through our newly created array, the loop lasting the length of the array. Inside that for loop we are going to check if our current maximum value, which starts at 0, less than the length of our current index of the array. If the max is greater, we skip it, but if the index of our array is greater, the length of that array index is our new maximum value.

Finally, when our loop is finished, we will return our maximum value.